+++ backend/src/main/java/com/festago/festival/domain/validator/school/ExistsFestivalSchoolDeleteValidator.java
@@ -1,9 +1,9 @@
-package com.festago.festival.application.validator.school;
+package com.festago.festival.domain.validator.school;
 
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
 import com.festago.festival.repository.FestivalRepository;
-import com.festago.school.application.validator.SchoolDeleteValidator;
+import com.festago.school.domain.validator.SchoolDeleteValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
