+++ backend/src/main/java/com/festago/stage/domain/validator/festival/ExistsStageFestivalDeleteValidator.java
@@ -1,8 +1,8 @@
-package com.festago.stage.application.validator.festival;
+package com.festago.stage.domain.validator.festival;
 
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
-import com.festago.festival.application.validator.FestivalDeleteValidator;
+import com.festago.festival.domain.validator.FestivalDeleteValidator;
 import com.festago.stage.repository.StageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
